refactor(db): extract VersionStore into src/server/stores/version-store.ts (#144)
Moves stash_versions / stash_version_files reads + restore into a
focused store, in the same style as the prior TokenStore / SessionStore
splits. Behaviour is bit-for-bit identical — pinned by the
db-versions.test.ts safety net merged in #161.

Surface:
- VersionStore.getStashVersions(stashId)
- VersionStore.getStashVersion(stashId, version, getStash)
- VersionStore.restoreStashVersion(stashId, version, getStash, by)
- VersionStore.insertVersionSnapshot({...}) — internal helper used by
  ClawStashDB.createStash + .updateStash inside their existing
  transactions; the version INSERTs now live next to the read paths.

ClawStashDB keeps its three public version methods as one-line
delegators so the existing route-handler call sites (versions/[v]/
restore + diff + list) do not change.

Numbers: src/server/db.ts 1549 -> 1400 LoC. 107 tests green, tsc clean,
build green.
